initscripts/patchtest: copy testing repo locally
For some reason, after a patchtest execution inside the guest machine
using the oe-core repo inside the share folder, the latter gets corrupted.
The corruption is observed by git when git status is executed inside the
host machine. It may be related to the 9p-virtio filesystem and/or the
machine being shutdown incorrectly, leading a dirty filesystem. One other
important finding that occurs often is that even if patchtest is not executed,
after the guest shutdown, this is observed in the host machine at the openembedded-core folder
shared with guest:

    diff --git a/meta/recipes-extended/ghostscript/ghostscript/i686 b/meta/recipes-extended/ghostscript/ghostscript/i686
    deleted file mode 120000
    index 87aaca5903..0000000000
    --- a/meta/recipes-extended/ghostscript/ghostscript/i686
    +++ /dev/null
    @@ -1 +0,0 @@
    -i586
    \ No newline at end of file
    diff --git a/meta/recipes-extended/ghostscript/ghostscript/i686 b/meta/recipes-extended/ghostscript/ghostscript/i686
    new file mode 100644
    index 0000000000..87aaca5903
    --- /dev/null
    +++ b/meta/recipes-extended/ghostscript/ghostscript/i686
    @@ -0,0 +1 @@
    +i586
    \ No newline at end of file
    diff --git a/scripts/pybootchartgui/pybootchartgui/main.py b/scripts/pybootchartgui/pybootchartgui/main.py
    deleted file mode 120000
    index b45ae0a3d2..0000000000
    --- a/scripts/pybootchartgui/pybootchartgui/main.py
    +++ /dev/null
    @@ -1 +0,0 @@
    -main.py.in
    \ No newline at end of file
    diff --git a/scripts/pybootchartgui/pybootchartgui/main.py b/scripts/pybootchartgui/pybootchartgui/main.py
    new file mode 100644
    index 0000000000..b45ae0a3d2
    --- /dev/null
    +++ b/scripts/pybootchartgui/pybootchartgui/main.py

These two files are the only softlinks found on the problem, so there is something going on
with the share folder and softlinks, which no root cause has been found.

The following change copies the openembedded-core repo locally and tests are done from the latter. Two other modifications:
increase the number of open files, and remove the explicit umount command which is not required. The former
was observed with several set of mboxes leading the following log:

    You can also run generated qemu images with a command like 'runqemu qemux86'
    /patchtest/mboxes/6627.4.mbox
    /patchtest/patchtest/scripts/test-mboxes: line 97: /patchtest/results/6627.4.mbox.log: Too many open files

[YOCTO #11707]

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>