patchtest: Fix handling of line endings
Patchtest wasn't applying some patches because of the line endings EDK2 uses.
If patchest used git am --keep-cr the patches would apply cleanly.

Also the Patch class in lib/patchestpatch.py in the patch test repo
reads the patch using f.read() with the newline parameter implicitly set
to None and this causes newlines to be translated from \r\n into \n.
Hence, when applying the patch with git am it fails; even when using
--keep-cr.

I was able to pass the patchtest tests with this change.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>